Check if known_hosts exists before trying to edit it

+++ lib/chef-vpc-toolkit/ssh_util.rb
@@ -5,6 +5,7 @@ module SshUtil
 	def self.remove_known_hosts_ip(ip, known_hosts_file=File.join(ENV['HOME'], ".ssh", "known_hosts"))
 
 		return if ip.nil? or ip.empty?
+		return if not FileTest.exist?(known_hosts_file)
 
 		existing=IO.read(known_hosts_file)
 		File.open(known_hosts_file, 'w') do |file|
